wined3d: Don't switch directly from one context to another (hack).
Switching from one context to another non-NULL context and then deleting
the original context triggers a bug in the Nvidia blob that causes a
crash on process cleanup. Other context changes on the same thread can
undo the damage. So usually we don't see the problem because the device
creation / destruction hides the problem.

The next patch however will eliminate the GL context for all threads
except the worker thread, uncovering this bug. Clearing the context
before activating the new one avoids the problem. This can be seen in
the ddraw overlay and d3d8/9 stateblock tests.

I must admit that I don't fully understand it yet. I haven't tried to
reproduce the issue in a stand-alone application yet, and for some
reason the crash occurs in the worker thread instead of the main thread,
but long after the worker thread has ended, and after THREAD_DETACH and
PROCESS_DETACH events have been processed. Lots of context activation /
deactivation activity in the worker can also hide the bug. In theory the
single threaded command stream should show this problem if an
application only creates a IDirect3D9 object without creating a device,
destroys it and then exits.