wined3d: Send palette updates through the command stream.
Right now the main purpose of this patch is to avoid a context
activation issue in the ddraw tests that was triggered by the
context_acquire call in the main thread. It is barely tested and missing
some pieces, like a synchronization concept that is better than just
waiting for the CS to finish.