Revert "winmm: Increase mmdevapi buffer sizes."
This reverts commit fdfff7cf6013a030f2b240b13b02831e502d49e7.

You've got to be kidding me, 1 second latency in winmm because
of a driver not used on linux? If the 'upstream' audio devs
spent less time on workarounds and more on fixing drivers,
this kind of crap wouldn't be needed.

I cannot believe that winecoreaudio requires .5 second latency,
their api allows for very low latency, so it's probably
winecoreaudio screwing things up instead.